Added the feature that the page is automatically refreshed after performing a CRUD operation

+++ src/ui/organisms/tables/TableServices.tsx
@@ -1,8 +1,9 @@
 "use client";
 import React from "react";
+import { useRouter } from "next/navigation";
+import { IGetServiceResponse } from "@/app/core/application/dto/services-salon/services-response.dto";
 import Table from "@/ui/molecules/common/table/Table";
 import PaginationServices from "../paginations/ServicesPagination";
-import { IGetServiceResponse } from "@/app/core/application/dto/services-salon/services-response.dto";
 import ButtonAdd from "@/ui/molecules/button-add/ButtonAdd";
 import Button from "@/ui/atoms/button/Button";
 
@@ -12,19 +13,25 @@ interface TableServicesProps {
 }
 
 const TableServices: React.FC<TableServicesProps> = ({ dataResponse, onEdit }) => {
+    const router = useRouter();
     const { content } = dataResponse;
 
     const handleDelete = async (id: number) => {
+        const isConfirmed = window.confirm("¿Estás seguro de que deseas eliminar este servicio?");
+        if (!isConfirmed) return;
+    
         try {
             await fetch(`/api/services/delete/${id}`, {
                 method: 'DELETE'
-            })
+            });
             console.log('Eliminado');
-
+            router.refresh();
+    
         } catch (error) {
             console.log('Error', error);
         }
     };
+    
 
     const formattedData = content.map((service) => ({
         name: service.name,
